fix: 21 bugs found by 3-agent audit (critical + high + medium)
CRITICAL FIXES:
- setup_data.py: skip-download → skip_download (runtime crash)
- dashboard.py: FEATURE_COLS 29→36, added velocity inputs + session_state
- dashboard.py: colors list 3→4 for baseline models
- dashboard.py: threshold 0.5→0.45 to match config
- api.py: Field(max_length=128) on transaction_id/user_id (DoS fix)
- features.py: avg_time_between_tx formula fixed
- features.py: int32→int64 for velocity counts
- .gitignore: .ipynyn_checkpoints → .ipynb_checkpoints
- run_baselines.py: broken imports fixed

HIGH FIXES:
- utils.py: guard duplicate logging handlers
- .dockerignore: created (was missing, bloated builds)
- docker-compose.yml: added models + config mounts to dashboard
- api.py: connected predict endpoint to feedback loop

MEDIUM FIXES:
- dashboard.py: removed dead groupby code
- dashboard.py: PR-AUC uses average_precision_score
- tests/test_features.py: rewritten with correct imports + 6 edge cases
- config.yaml: added model.path key
- README.md: fixed metrics.json path

All 9/9 stress tests passed.

+++ src/features.py
@@ -176,7 +176,7 @@ def _create_rolling_velocity_features(df: pd.DataFrame) -> pd.DataFrame:
             df[col_sum] = 0.0
             df[col_mean] = 0.0
 
-        df[col_count] = df[col_count].fillna(0).astype(np.int32)
+        df[col_count] = df[col_count].fillna(0).astype(np.int64)
         df[col_sum] = df[col_sum].fillna(0.0)
         df[col_mean] = df[col_mean].fillna(0.0)
 
@@ -189,7 +189,7 @@ def _create_rolling_velocity_features(df: pd.DataFrame) -> pd.DataFrame:
         df = df.merge(
             all_prev[["prev_count", "last_step"]], on="_orig_idx", how="left"
         )
-        df["prev_count"] = df["prev_count"].fillna(0).astype(np.int32)
+        df["prev_count"] = df["prev_count"].fillna(0).astype(np.int64)
         df["last_step"] = df["last_step"].fillna(0)
 
         df["avg_time_between_tx"] = np.where(
